requirements.txt
streamlit
ultralytics
torch
pandas
Pillow
